% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{bees.mp4}
\alias{bees.mp4}
\title{Video of "Bee on a flower"}
\description{
The original video of bees.mp4 was downloaded from \url{https://mixkit.co/free-stock-video/bee-flower-1000/}.
We can use the original video of bees.mp4 under Mixkit license. See \url{https://mixkit.co/license/} for detail.
bees.mp4 was made by trimming and cropping the original video of bees.mp4 to reduce the file size.
}
\examples{
system.file("extdata", "bees.mp4", package = "videor")
}
