% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{bees}
\alias{bees}
\alias{bees.avi}
\alias{bees.mp4}
\alias{bees.webm}
\alias{bees.wmv}
\title{Video of "Bee on a flower"}
\description{
The original video of bees was downloaded from \url{https://mixkit.co/free-stock-video/bee-flower-1000/}. \cr
We can use the original video of bees under Mixkit license. See \url{https://mixkit.co/license/} for detail. \cr
bees.mp4 was made by trimming and cropping the original video of bees.mp4 to reduce the file size.
bees.avi, bees.webm, and bees.wmv were converted from bees.mp4.
}
\examples{
system.file("extdata", "bees.avi", package = "videor")
system.file("extdata", "bees.mp4", package = "videor")
system.file("extdata", "bees.webm", package = "videor")
system.file("extdata", "bees.wmv", package = "videor")
}
