% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_id.R
\name{clean_id}
\alias{clean_id}
\title{Helper Function to clean ids fed to other functions}
\usage{
clean_id(ids)
}
\arguments{
\item{ids}{character vector}
}
\value{
ids character vector
}
\description{
Takes in vector of ids to remove NAs and whitespace.
}
\examples{
clean_id(c(" 12hi5nh ", "12rfe34rt", NA))

}
