% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_req.R
\name{get_req}
\alias{get_req}
\title{Helper Function to send API requests}
\usage{
get_req(url, token)
}
\arguments{
\item{url}{character}

\item{token}{environment}
}
\value{
temp data.frame
}
\description{
Takes in custom url and token to get request
}
\examples{
get_req("https://example.com", token_var)

}
