% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytanalytics_country_stats.R
\name{ytanalytics_country_stats}
\alias{ytanalytics_country_stats}
\title{YouTube Analytics API Wrapper}
\usage{
ytanalytics_country_stats(
  token = NULL,
  maxResults = 10,
  startDate = Sys.Date() - 30,
  endDate = Sys.Date()
)
}
\arguments{
\item{token}{environment}

\item{maxResults}{numeric}

\item{startDate}{character string}

\item{endDate}{character string}
}
\value{
df data.frame
}
\description{
This function returns number of Views, Shares, Likes, Dislikes, Comments,
Subs Gained & Lost for the channel by country (Can adjust date range)
}
\examples{
ytanalytics_country_stats(token_var, 10, "2020-05-01", "2020-05-31")

}
