% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytanalytics_request.R
\name{ytanalytics_request}
\alias{ytanalytics_request}
\title{YouTube Analytics API Wrapper}
\usage{
ytanalytics_request(
  dimensions = NULL,
  metrics,
  sort,
  maxResults = 10,
  filtr = NULL,
  startDate = Sys.Date() - 30,
  endDate = Sys.Date(),
  token,
  channelId = "MINE"
)
}
\arguments{
\item{dimensions}{character string}

\item{metrics}{character string}

\item{sort}{character string}

\item{maxResults}{numeric}

\item{filtr}{character string}

\item{startDate}{character string}

\item{endDate}{character string}

\item{token}{environment}

\item{channelId}{character string}
}
\value{
df data.frame
}
\description{
This function is an API helper function for pulling
YT Analytics data on your own channel
}
\examples{
ytanalytics_request("video", "views", "-views", 10, NULL, "2020-05-01", 2020-05-31, token_var, "MINE")

}
