% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytanalytics_sharing_service.R
\name{ytanalytics_sharing_service}
\alias{ytanalytics_sharing_service}
\title{YouTube Analytics API Wrapper}
\usage{
ytanalytics_sharing_service(
  token = NULL,
  maxResults = 10,
  startDate = Sys.Date() - 30,
  endDate = Sys.Date()
)
}
\arguments{
\item{token}{environment}

\item{maxResults}{numeric}

\item{startDate}{character string}

\item{endDate}{character string}
}
\value{
df data.frame
}
\description{
This function returns number of Shares by sharing service
}
\examples{
ytanalytics_sharing_service(token_var, 10, "2020-05-01", "2020-05-31")

}
