% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytanalytics_top_videos.R
\name{ytanalytics_top_videos}
\alias{ytanalytics_top_videos}
\title{YouTube Analytics API Wrapper}
\usage{
ytanalytics_top_videos(
  token = NULL,
  maxResults = 10,
  startDate = Sys.Date() - 30,
  endDate = Sys.Date()
)
}
\arguments{
\item{token}{environment}

\item{maxResults}{numeric}

\item{startDate}{character string}

\item{endDate}{character string}
}
\value{
df data.frame
}
\description{
This function returns number of Views, Likes, Dislikes, and Comments
on top videos (Can adjust date range and max results (from 1-200))
}
\examples{
ytanalytics_top_videos(token_var, 10, "2020-05-01", "2020-05-31")

}
