% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytanalytics_traffic_sources.R
\name{ytanalytics_traffic_sources}
\alias{ytanalytics_traffic_sources}
\title{YouTube Analytics API Wrapper}
\usage{
ytanalytics_traffic_sources(
  token,
  maxResults = 10,
  startDate = Sys.Date() - 30,
  endDate = Sys.Date()
)
}
\arguments{
\item{token}{environment}

\item{maxResults}{numeric}

\item{startDate}{character string}

\item{endDate}{character string}
}
\value{
df data.frame
}
\description{
This function returns the views by traffic type
}
\examples{
ytanalytics_traffic_sources(token_var, 10, "2020-05-01", "2020-05-31")

}
