% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_channel_most_recent_video.R
\name{ytdata_channel_most_recent_video}
\alias{ytdata_channel_most_recent_video}
\title{YouTube Data API Channel Wrapper}
\usage{
ytdata_channel_most_recent_video(
  channelId = NULL,
  token = NULL,
  maxResults = 1
)
}
\arguments{
\item{channelId}{character string}

\item{token}{environment}

\item{maxResults}{numeric}
}
\value{
df data.frame
}
\description{
This function pulls a channel's most recent video(s)
}
\examples{
ytdata_channel_most_recent_video("as09df8yaidfjkl", token_var)

}
