% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_channel_playlists.R
\name{ytdata_channel_playlists}
\alias{ytdata_channel_playlists}
\title{YouTube Data API Channel Wrapper}
\usage{
ytdata_channel_playlists(channelId = NULL, token = NULL)
}
\arguments{
\item{channelId}{character string}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function returns ALL playlists in a channel and their Published Date/Time,
Title, Description, Channel Title & ID, Video Count and Playlist ID
}
\examples{
ytdata_channel_playlists("as09df8yaidfjkl", token_var)

}
