% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_channel_stats.R
\name{ytdata_channel_stats}
\alias{ytdata_channel_stats}
\title{YouTube Data API Channel Wrapper}
\usage{
ytdata_channel_stats(channelId = NULL, token = NULL)
}
\arguments{
\item{channelId}{character string}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function returns Title, Description, CustomURL (if applicable),
Published Date/Time, Country, and number of Views, Comments, Subscribers,
and Videos.
}
\examples{
ytdata_channel_stats("as09df8yaidfjkl", token_var)

}
