% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_channel_videos.R
\name{ytdata_channel_videos}
\alias{ytdata_channel_videos}
\title{YouTube Data API Channel Wrapper}
\usage{
ytdata_channel_videos(channelId = NULL, token = NULL)
}
\arguments{
\item{channelId}{character string}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function loops through each video in each playlist to get all
channel videos and their stats
}
\examples{
ytdata_channel_videos("as09df8yaidfjkl", token_var)

}
