% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_playlist_videos.R
\name{ytdata_playlist_videos}
\alias{ytdata_playlist_videos}
\title{YouTube Data API Playlist Wrapper}
\usage{
ytdata_playlist_videos(playlistId = NULL, token = NULL)
}
\arguments{
\item{playlistId}{character string}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function returns list of videos within a playlist
}
\examples{
ytdata_playlist_videos("as09df8yaidfjkl", token_var)

}
