% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_search.R
\name{ytdata_search}
\alias{ytdata_search}
\title{YouTube Data API Search Wrapper}
\usage{
ytdata_search(term = NULL, token = NULL, maxResults = 10)
}
\arguments{
\item{term}{character string}

\item{token}{environment}

\item{maxResults}{numeric}
}
\value{
df data.frame
}
\description{
This function returns Publish Date/Time, Channel ID, Title, Description,
Channel Title, and LiveBroadcastContent (if Applicable) for search term
}
\examples{
ytdata_search("Sail Cat", token_var)

}
