% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_video_details.R
\name{ytdata_video_details}
\alias{ytdata_video_details}
\title{YouTube Data API Video Wrapper}
\usage{
ytdata_video_details(ids = NULL, token = NULL)
}
\arguments{
\item{ids}{character vector}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function returns PlayTime, Dimension(?), Definition (HD, etc),
Captions (bool), and License (bool)
}
\examples{
ytdata_video_details("21ljkho84", token_var)

}
