% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_video_metadata.R
\name{ytdata_video_metadata}
\alias{ytdata_video_metadata}
\title{YouTube Data API Video Wrapper}
\usage{
ytdata_video_metadata(ids = NULL, token = NULL)
}
\arguments{
\item{ids}{character vector}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
This function returns Publish Date/Time, Channel ID, Title, Description, Tags,
Category ID, LiveBroadcastContent (if applicable), and DefaultAudioLanguage
}
\examples{
ytdata_video_metadata("21ljkho84", token_var)

}
