% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata_video_stats.R
\name{ytdata_video_stats}
\alias{ytdata_video_stats}
\title{YouTube Data API Video Wrapper}
\usage{
ytdata_video_stats(ids = NULL, token = NULL)
}
\arguments{
\item{ids}{character vector}

\item{token}{environment}
}
\value{
df data.frame
}
\description{
Takes in vector of ids and a token to get overall number of Views,
Likes, Dislikes, Favorites, and Comments of each video
}
\examples{
ytdata_video_stats("21ljkho84", token_var)

}
