% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{bees.mp4}
\alias{bees.mp4}
\title{Video of "Bee on a flower"}
\source{
\url{https://mixkit.co/free-stock-video/bee-flower-1000/}
}
\description{
This video was downloaded from \url{https://mixkit.co/free-stock-video/bee-flower-1000/}.
We can use this video under Mixkit license. See \url{https://mixkit.co/license/} for detail.
}
\examples{
system.file("extdata", "bees.mp4", package = "videor")
}
