% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_video.R
\name{info_video}
\alias{info_video}
\title{Video Information}
\usage{
info_video(video)
}
\arguments{
\item{video}{path to a video file}
}
\value{
a list of video information
}
\description{
get video information such as width, height, format, duration and framerate.
Note that this function is relatively slow except when video is mp4 file.
}
\examples{
video <- system.file("extdata", "bees.mp4", package = "videor")
info_video(video)
}
\author{
Shota Ochi
}
