% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_video.R
\name{load_video}
\alias{load_video}
\title{Loading Video}
\usage{
load_video(video, frames)
}
\arguments{
\item{video}{path to a video file}

\item{frames}{a numeric vecotr that specfies a frame or a numeric vector that specifies start frame and end frame}
}
\value{
a list of video information
}
\description{
load specified frames of video
}
\author{
Shota Ochi
}
