% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_video.R
\name{load_video}
\alias{load_video}
\title{Loading Video}
\usage{
load_video(video, frames, nb_frames)
}
\arguments{
\item{video}{path to a video file}

\item{frames}{a numeric vecotr that specfies a frame or a numeric vector that specifies start frame and end frame. if frames is missing, try to load all frames.}

\item{nb_frames}{the number of frames. if nb_frames is missing, info_video(video)$frames is used.}
}
\value{
a cimg object
}
\description{
load specified frames of video
}
\author{
Shota Ochi
}
